sdk: go: move querybuilder to internal
