+++ codegen/generator/templates/src/header.go.tmpl
@@ -6,7 +6,7 @@ import (
 	"context"
 
 	"github.com/Khan/genqlient/graphql"
-	"dagger.io/dagger/querybuilder"
+	"dagger.io/dagger/internal/querybuilder"
 )
 
 // graphqlMarshaller is an interface for marshalling an object into GraphQL.
