+++ sdk/go/client.go
@@ -9,7 +9,7 @@ import (
 	"dagger.io/dagger/internal/engineconn"
 	_ "dagger.io/dagger/internal/engineconn/embedded" // embedded connection
 	_ "dagger.io/dagger/internal/engineconn/unix"     // unix connection
-	"dagger.io/dagger/querybuilder"
+	"dagger.io/dagger/internal/querybuilder"
 	"github.com/Khan/genqlient/graphql"
 	"github.com/vektah/gqlparser/v2/gqlerror"
 )
