+++ sdk/go/internal/querybuilder/marshal.go
