+++ sdk/go/internal/querybuilder/querybuilder.go
