+++ sdk/go/internal/querybuilder/querybuilder_test.go
